update ocaml executables to conform to new search-in-advance paradigm
I saw in earlier experiments that gradient-based optimization was struggling
to work with compression to produce a genuine diversity of discrete representations.
I decided to redesign to cut out this issue entirely by executing random programs
ahead of time and saving their results.  The bottlenecked codebook will consist of
vectors the size of an entire entry in a batch, and each will be mapping to one of
the distinct program outputs that resulted from the intermittent random searching.
I call this procedure exploration, but in DreamCoder parlance it would be called
dream enumeration/dream sleep.

This design better supports the original goal, of having the PSN choose between
equivalence classes of programs (where equivalence is defined by their output
graph).  It also better demonstrates the general paradigm of PSN training, and
creates a more flexible separation between the program synthesis element and the
neural network design---both of which are necessary, and probably permitting their
own independent improvements through future work.