checkpoint code before full shift to low-data regime training
Previously, we trained on very small chunks of the full dataset
by cycling through them incrementally.  I am now investigating what
happens if we don't cycle at all (that is, essentially train from
<100 puzzles).  It is easiest to do this if I entirely remove the
routines used for cyclic training, as in practice I was not running
them for long enough for the cyclicality to matter.  It was essentially
a small data regime but with more complexity required to limit the
pool of training data.  This commit captures the following changes:

 * switching to a search-in-advance paradigm in which the `in_codebook`
   vectors are the size of the entire pre_quantizer output, and correspond
   to distinct graphs computed ahead of time.  Thanks to this change, we
   could remove the "contextual" functionality from the restart manager.
 * always return the restart vectors in descending order of time since
   last use and never return more restartable vectors than you received
   quantizations (as restarting more than this number of vectors could
   mean you are actuallly setting some codebook vectors to the same value).
 * fix issue with setting of `out_restarts` in `PSN` that was accumulating
   more restarts than their were pre_quantizer outputs to restart them
   with
 * add method to `PSN` class for performing exploration (search-in-advance)
 * add methods to `RavenDataset` and `train.py` for training with a single
   small training set and large evaluation set.