adapting ocaml entry points for use of optimal distributional search and stitch compression tool
I have conducted low-data regime training through a number of methods.
The major question that arose during this process was exactly how
search, compression, and PSN training should interact.  In recent
commits, I pivoted to a search in advance paradigm; this implied that
programs were not synthesized and executed during PSN inference;
rather, program search was conducted in advance to develop sets
of I/O examples (graphs, in our case), and the PSN chose among vectors
corresponding diectly to these outputs.  At the time, I understood
this change as removing all influence of PSN training on program
search.  Initially, I conceived that program search would be done
via fixed procedure, and compression would be used to iteratively
compress this search space, allowing a usefully large and complex
set of graphs to be found.  In experimenting with this, I relearned
the importance of one of the major techniques from DreamCoder:
distributional search.  While compression helps program search to
reach novel territory by trading search depth for search breadth,
this can only do so much to ameliorate the inherent difficulty of
searching such a large space.  Without some form of bias toward
particular regions of the search space, search will remain shallow,
eventually leading to a stagnation in the degree of novelty that
can be achieved via compression.  For the reachable search space
to continue to grow in sophistication, it must become deeper and
less broad.  Otherwise, growth in the pool of novel I/O examples
found will pleateau.

After a number of iterations of non-distributional search
implementations followed by implementation of DreamCoder-style search,
I saw that recent papers by Ellis and company had long since innovated
on these two basic elements of DreamCoder with optimal forms of
distributional search (heap search, SQRT search) and more scalable
non-deductive forms of compression (Stitch).  This meant that the
only unique function of the OCaml codebase was in stitching these
pieces together into an exploration-compression system.  Since a
good bit of design had gone into this aspect of the OCaml codebase
and neither of the more recent distributional search/compression papers
were actually made to interoperate as a singular tool for exploration-
compression, I decided to reimplement heap search in OCaml--replacing
the DreamCoder-style search implementation--and wrote glue to incorporate
the results of stitch compression in a manner compatible with the OCaml
DreamCoder-style implementation of compression.

Changes include:
 - slight modifications to the parameters of the DreamCoder-style
   compression procedure, mainly to incorporate the capability to restrict
   the set of invention candidates based on size (choosing invented
   primitives only from those with one of the `n_invention_sizes` smallest
   sizes amongst the possible candidates).  This function was an
   experimental concept from the trials of uninformed search-in-advance.
 - reestimation of DSL parameters prior to program search based on a
   frontier of programs producing graphs currently favored by the PSN.
 - addition of an `incorporate_stitch` script and corresponding additions
   of DSL-sensitive parsing capabilities to facilitate reading stitch-rewritten
   programs containing novel primitives referenced by name only (rather than
   anonymously as inline implementations).