debug new exploration and compression capabilities
To assimilate new stitch primitives, `incorporate_stitch` needed to be
able to parse invented primitive bodies that followed stitch syntax in
which the parameters of a function are labeled in reverse application
order and prefixed with "#" rather than "$".  This required additional
parsing functionalities to be added to `antireduce` and utilized in the
OCaml scripts.  The following bugs were also fixed:
 - The argument of `GraphQuantizer._load_representations` was expected
   as a dict, but was not a dict at the calling locations.  Addressed by
   calling dict before passing argument to `_load_representations`.
 - The `disable_stat_logging` option documented as being a possible
   stitch keyword-argument raised an error when supplied as true; fixed
   by no longer supplying this argument.
 - While the `dreamcoder_compress` script expects the frontier as a list
   of filenames, `stitch_core.compress` expects a list of program strings.
   Fixed by loading the `stitch_program` entry of the representations JSON
   for each filename in `frontier` before calling `stitch_core.compress`.