refactor training routines
Refactored training routines to enforce nn-training -> compression ->
program search cycle.  This was one of a number of possible cycles that
was possible before; with this restriction, PSN training with program
synthesis maintains the same basic order of operations as used in DreamCoder.

In restucturing these routines within `experiment.py`, `psn.py`, and
`quantizer.py`, I favored local functions accumulating complete logs and
returning JSON (dicts) of information about the calls that they performed.
This makes it easier for the high-level training routines to accumulate
logs, since information about parameters is logged at site of use and
accumulated in a bottom up fashion from whatever calls are involved.
While this increases duplication of parameter logging provided those
parameters are being used consistently, as expected, they also serve as
a safety check to document the consequences if this is somehow not the case.
So the duplication is not without its benefits.

I also observed that if the frontier was in fact being probabilistically
sampled, it would be problematic that this sampling was done twice, once
in the `PSN.compression` subroutine and once in the `PSN.exploration`
subroutine.  We want the same frontier to be used for both, even if some
randomness was utilized to create it.

I also saw that the unused and irrelevant `program_size` field in the
`__init__` function of the `PSN` class was playing a role in determining
which forward inference modes could be used by determining how the in
and out codebooks were instantiated.  This mechanism was misleading as
program search no longer adheres to a strict program size limit; whether
or not this was provided actually indicated whether a model was a PSN
utilizing a dual set of codebooks with some form of quantizer between
them, or whether it was a non-quantizing or single-codebook quantizing
equivalent to a PSN utilizing equivalent subnetworks.  This distinction
was made clearer, and the instantiation of the second codebook is now
prevented in the case that it will not/cannot be used.  The distinction
is tracked and used when validating the `mode` parameter of `PSN.forward`:
models that were initialized to use two-stage quantization cannot operate
in the "none" or "nearest" modes, only in "learned" mode.  Models not
initialized to run in two-stage quantization mode can run in "none" or
"nearest" mode, not in "learned."

minor field renaming of some JSON fields associated with `GraphQuantizer.explore`
and `GraphQuantizer.dreamcoder_compress` was performed; a new method was
added to the `GraphQuantizer` class to capture a more flexible ability
to form numerically-incremented DSL names from a base string.