conform to updates in OCaml dependencies; make dsl-insensitive parsing not require JSON
The distinction between DSL-sensitive and DSL-insensitive parsing was not
genuine, as the domain generic forms of these functions allowed parametrization
with arbitrary JSON.  This was done to acommodate a very `antireduce-graphs`
specific use case: that depending on the value of `max_color` different
primitives were (at one time) presumed to be present in the DSL.  This was
an unhygienic workaround; it would make more sense (or at least create
clearer semantics) to have neither of these functions be parametrized by
JSON; the dsl-insensitive version would then be parametrized only by its
domain, and the dsl-sensitive version would be parametrized by its domain
and a DSL.  Variation in the DSL triggered by a choice of a parameter in
Python which affects parsing would then have to be factored in at the
barrier between the two languages, rather than being assumed both within
the OCaml code implementing a domain and also within relevant Python code.

The OCaml codebase was significantly modularized as well; these namespace
changes required minor updating.