refactor to abide by assumption of uniformly course two-stage quantization
As described in the previous commit, I am changing the PSN design to assume
that two-stage quantization is always performed with two codebooks of equal
dimensionalities whose first dimensions are equal to the number of discrete
representations recognized by the quantizer, and whose second dimensions are
equal to `seq_length * seq_dim` for the sequence length and dimension used
by the ViT modules surrounding it.  With this change, benchmarking against
the VQ-VAE, or either of the previously-designed "bottleneck quantizer"
ablations is no longer informative, as none of the relevant designs are
particularly interesting or well-optimized when constrained to utilize
codebook(s) of this shape.  As a result, it is no longer important to maintain
the multiple "modes" of the PSN class as previously designed, nor to maintain
a separation between a quantizer abstraction and a PSN abstraction.  Both
may simply be collapsed into a parent `Quantizer` class (now residing in
its own `quantizer` module) and an inheriting `GraphQuantizer` class which
implements a much more constrained set of methods.  This new quantizer class
melds the `exploration`/`compression` and `forward` methods of the former
`PSN` class and `GraphQuantizer` classes into generic methods of the parent
`Quantizer` class, leaving only the "reconstruction" transformation from
`in_codebook` encoding indices back to `codebook_dim`-sized vectors to be
implemented by child classes.  Two utility functions `clear_representations`
and `load_dsl` have also been added, allowing the state of the program synthesis
backend to be fully reset using methods of the new `Quantizer` class.

These simplifications allow a very simple class `ExperimentalModel` within
the `experiments` module to implement the baseline (quantization-free) model
and a PSN containing a `GraphQuantizer`.  This model implements all of
the running/training procedures that previously floated freely in the
`experiments` module as methods.  This allows the implementation of a single
`train_split` method whose behavior depends on the presence of an intermediate
quantization module.  If present, it will perform the fully DreamCoder-style
program synthesis cycle, and if not will perform only the neural network
training component of this procedure.

The various parameters previously represented as constants of the
`experiments` module such as batch size, initial dsl name, and others
have been converted to true parameters of the relevant methods to better
facilitate experimentation from within the Python interpreter.

The long-since unused `request.json` file is also removed with this
commit.

The strided-convolution stem ViT models, plain ViT models, unused parameters
of pixelshuffle-stem ViT models included to retain backward compatibility,
and the positional embedding layers used in all of the above to encode the
effects of fine-grained quantization have all been removed from the `networks`
module, as these variations have long proven unimportant to the experiments at
hand.