minor bug fixes
So far, the removal of fine-grained quantization via `out_codebook` does
not seem to have changed the plateauing accuracies I have experienced up
to this point.

In my previous draft of the paper to accompany this code, I was very
focused on the fact that in choosing between courser and finer-grained
nearest-neighbor quantization, one was essentially choosing between more
flexible discrete structure and more flexible continuous structure in
the representation.  I think my underlying intuition was that neural
networks were probably not making much use of the discrete structure
that was created by nearest-neighbor quantization.  Indeed, the
dimensionalities typically used for VQ-VAE-descended architectures
suggest that the goal is generally to be so fine that the discrete
information available is too subtle and voluminous to be of much use.

I believed that if one initially used a courser quantization, then
mapped this choice onto a more complex discrete representation from
which the original continuous information must be discerned with
increased greater precision, I would force the neural network to reckon
with the role of discrete structure of in the more finely quantized
representation.  This has occurred when the prequantizer and GNN have
independently learned to predict the same thing on a different basis.
The out-codebook vectors are gradually made to resemble prequantizer
outputs, and therefore the training data, and the GNN is encouraged
to closely reconstruct these vectors purely from the discrete structure
corresponding to in-codebook vectors.  Consequently, the entire model
should succeed when there is a profitable alignment between training
data and discrete representations that allows out-codebook
representations and discrete representations to coincide.  When this
occurs, the resulting model should generalize more quickly and easily,
as there are fewer "bins" with which to identify new data according
to the course quantization, even if its representation subsequently
becomes more finely differentiated.

This theory presents a compelling case: by constraining the realizable
outputs of the model, course quantization makes it easier to generalize
at test time but harder to learn.  We can ameliorate this by making the
categorical choices of representation the network faces more structured,
informative, and distinct.  However, I don't think the mechanism it
lays out accomplishes this.  As described, the PSN architecture simply
institutes a requirement that the model be able to map the discrete
information captured by a categorical choice of in-codebook vector
(a graph representing its index) to an approximation of the continuous
information captured therein.  In an autoencoding task where training
where there is a clear incentive to map each training data point to a
distinct output, there is no risk of codebook collapse, and training
data outnumber discrete representations, this constraint would force
the model to consider the discrete structure of each of its (limited)
choices.  However in any less perfect condition, such as an autoencoding
task with limited training data or a more typical classification task,
there is little reason to think that the categorical choices faced by
the network are made more meaningfully distinct by this mechanism; it
seems more likely that the potentially large approximation in copied
gradients only encourages them to be less so at convergence.

A better mechanism would ensure that the representation resulting from
the categorical choice of in-codebook vector is always enhanced with
discrete structure through a novel join representation.  Without the
need to enforce equivalence between two separate representations there
is no need to isolate the GNN, and consequently no need for two-stage
quantization.  Thus we may backpropagate through the network in a much
more ordinary way, only using stopgradients for the single course
quantization, as in the VQ-VAE.

In particular, we may design a kind of residual unit where the coursely
quantization of the prequantizer output is packed into the node
representation, and the output of the GNN is transformed such that it
can be added back to the original quantized output before processing
by the postquantizer.  This residual unit with single-stage quantization
still functions as a generic neural network component, but arguably
in a more compelling way.  Furthermore, because the GNN has some context
for how to interpret a particular graph, and is also graded according
to its success at the actual downstream prediction task, we have every
reason to think that the feedback given to the prequantizer bears on
the relevance of the singular codebook's continuous and discrete structure.
Consequently, the prequantizer's choice of codebook index should also
be more informative in guiding distributional program search.

It is worth noting that this scheme is similar in spirit to noisy
generalization.  However, rather than imposing a very simple,
"one dimensional" kind of adjacency between neighboring codebook
vectors, these adjacencies are instead defined by more general
relational structure.

I am hoping that this altered design of a Program Synthesizing Residual
Unit will achieve better than the plateauing accuracies I have continued
to experience with the PSN

In anticipation of the aforementioned changes, a number of bug fixes are
being committed so as to checkpoint the most recent PSN design.  These
include:

 - fix various broken calls from API changes
 - enforce that by passing an empty frontier to `GraphQuantizer.explore`
   you are specifically requesting uniform distributional search
   (i.e. breadth-first search)
 - the mechanism for selecting 2d positional embeddings was not functioning
   properly, as it appears the previous attempt at multidimensional indexing
   was not working as expected; this was fixed by using calls to `torch.gather`.