rigorously measure baseline
I have rigorously measured the baseline and had some results that
I think are promising with the PSN.  I have further modified the
design so that it is not a residual unit, but one in which all
backpropagation must occur via the GNN.

In trying to run my model 10 times, I observed that the performance
was dramatically slowing because the codebook was becoming so large
(2000+ vectors in one case).  I realized that of course this isbad
from a performance perspective, but it's terrible for the theory of
how/why the PSN is supposed to work.  If there's so many discrete
representations that you could assign one to each training data
point multiple times over, then you're hardly instituting a bottleneck
at all.  Particularly bad: the test data could be quantizing to
representations that there isn't even enough training data to vet.
The entire purpose is that in performing inference at test time,
you are simply binning data into bins that have already proven
to be more general through experience with the training data.  I
believe that for this to occur, the number of representations has
to be less than the number of training data points.  Of course,
it must also be larger than the number of possible class labels.
In general usage, those two numbers might be very difficult to
precisely state; one would just have to play around with them.
Nonetheless, a more robust approach should be taken than simply
retaining every representation ever discovered.  Unless we are
more selective than this, there is no way to make sure we are
creating a bottleneck, much less one whose discrete structure is
constantly evolving in challenging ways.  I believe that even as
the programs get more sophisticated in the current regime, because
all programs are retained, this doesn't really create greater
challenge for the network; just a bit more temporary busywork.

In order to make room for new representations to still be introduced
as new representations are discovered, we need to cull the representations
of some percent of its least frequently selected members.  What remains
will be the frontier, which will then be refilled with whatever number
of new representations can be discovered within a certain time limit
and do not exceed the given maximum size.  The changes saved with
this commit include:

 - correction of unequal parameters between `raven_baseline` and
   `raven_psn`
 - addition of vqvae-style baseline testing the influence of
   quantization alone.
 - modification of semantics of `beta` parameter; it now shrinks the
   prequantizer loss by a factor of beta if beta is a fraction less
   than one, and shrinks the codebook loss by a factor of 1 / beta
   if beta is greater than one.
 - implementation of a modification of the `Quantizer` class so that
   it is no longer a residual unit; all backpropagation occurs via
   the GNN.