limit number of discrete representations to less than number of training data points
This change appears to be having a positive effect; however, it has
highlighted that with this initial DSL, program synthesis can lead to
some highly homogenous pools of graphs.  This creates a tradeoff: I
suspect that generally, the fewer representations used, the more
quickly and easily generalization should occur.  However, this also
gives program synthesis less opportunity to produce a diversity of
graphs.  If, with a given initial DSL, program synthesis is not very
efficient at producing designs that offer genuinely different sets
of constraints for neural network influence, it becomes very difficult
to tell exactly how many discrete representations one should allow.
I suspect that these "sets of constraints" offered by utilizing a
GNN embedding with a particular codebook vector, are more "complex"
when the graphs are more densely connected, and offer more a more
uncomplicated opportunity to reconstruct the original codebook vector
when they are sparsely connected.

This change is entirely implemented from OCaml.  Nonetheless, I wanted
to checkpoint the changes made so far.  I suspect that most of the
remaining design effort will indeed be outside of this particular
repository (except to contribute utilities for running complete
experiments more easily).