differentiate number of representations used during evaluation from number used while learning
After observing anecdotal results following the changes of the previous
commit, I came to believe that there's reason to think that the variety
of discrete representations necessary/beneficial for learning in a PSN
may be different from the number of discrete representations necessary/
beneficial for inference.  That is, a greater variety may be helpful
when learning which such structures are most relevant to the prediction
task at hand, but may be harmful when encountering out-of-distribution
data.  In the latter case, it could be helpful to be required to stick
to a limited subset of structures that have been found most useful for
characterizing the training data.

It is still desirable to have some turnover in discrete representations,
so that those which are highly-infrequently used don't stick around
forever, and so that compression may have its desired effect of raising
the overall structural complexity of the body of discrete representations
over time.  Nonetheless, the size of this body and the turnover it
experiences provide a rather crude way of limiting the representations
available during inference.  We may retain the previously-built mechanism
allowing turnover of some fraction of least-used discrete representations
and simply build a separate mechanism for segregating a fraction of the
most highly-used and most "archetypal" representations.  This also allows
us to consider space/program synthesis time constraints on our desired
training diversity apart from our desired inference diversity.  The
previous behavior is simply the restricted (and still accessible) case
in which the number of top-performing discrete representations preserved
for training diversity and the number utilized during inference are the
same.

This design required/accompanied a number of changes:
 - creation of a number of new core routines in `quantizer.py`
   that enable a smaller "archetypal" codebook to be used during
   inference and a larger "base" codebook to be used in training.
 - internalization within the quantizer of the tracking previously
   done externally, via `ExperimentalModel._run` with `repr_usage`.
   Since counting the usage of particular representations is now
   necessary to maintaining clearly `Quantizer` internal data such
   as its codebooks, relied upon for the implementation of `forward`,
   it makes more sense simply to maintain these counts internally
   in manner that can also be relied upon externally for logging.
   Thus, rather than methods of `ExperimentalModel` recording
   representations usages themselves and passing this information
   between methods to inform exploration, compression, and inference,
   equivalent but better-maintained information is maintained by
   its `quantizer` and is simply accessed at appropriate times by
   its containing class.  This makes certain new parameters such
   as `usage_smoothing_factor`, `n_archetypes`, and `n_preserved`
   relevant to `ExperimentalModel._run` contingent on the presence
   of a quantizer.  The previous practice of logging closest to
   proximal usage at the expense of being verbose/redundant was
   continued and all methods contingent on `ExperimentalModel._run`
   were refactored.
 - within `graph_quantizer.py`, modifications were made to prevent
   bidirectionally-duplicated edges from having different features
   (the duplicated edge in both directions should have the same edge
   creation order ID).  Additionally, dropout was added between all
   layers of the GNN.

Sadly, none of the above implemented changes seemed to produce the
desired performance, which for the first time was measured rigorously
as an average of 10 runs.  Nonetheless, upon doing so I noticed that
across the iterations the PSN was displaying a unique pattern: its
performance would spike in earlier iterations then get progressively
worse.  Whereas the baseline, which in previous examinations I had
accidentally summed across entire runs rather than across iterations,
did not display this pattern, exhibiting consistency across iterations
or perhaps a slight increase in evaluation performance in later
iterations.  This gave me a suspicion that perhaps the form of
codebook vector-graph embedding was to blame.

Ideally, as compression, combined with consistent turnover, forces
graphs to grow more complex, we would like this complexity to be
increasingly influential w.r.t. how the codebook vector embedded on
it is processed.  We would like it to have a greater role in structuring
the network's inference.  However, in the previous scheme the codebook
vector was simply replicated across all nodes of the graph, with
dropout applied uniformly.  This essentially meant that the larger
a graph was, the more partially-redacted copies of the codebook vector
it corresponded to became available to the GNN--and the more this GNN
could succeed by setting no goal but to reconstruct this vector from
the original copies.  Or, that is, experience temporary success.
Because as graphs got larger, the graph's structure would inevitably
have more influence on the GNN's output.  However, due to the randomness
and inconsistency of dropout, and the random resetting of the codebook
vectors themselves, one would think that the results of this influence
would grow increasingly difficult to learn from and control.

The solution is one that I had thought about in the past in some form
already: embed pieces of the codebook vector throughout the graph
so that the way they are distributed onto the graph structure is
not random but reflective of how it was created by the original
program.  There are, presumably, many ways to do this.  Ideally,
I think we would want to choose one such that the graph neural
network doesn't gain anything in particular by being globally-aware
of the number of nodes in the graph.  That is, we want each neighbor-
hood of nodes to "stand on its own" in terms of the amount of detail
it contains.  I think the best, and most easily size-invariant
technique for doing this is to divide the codebook vector into
`n_elts` chunks, and tile each of these chunks into a `graph_dim`
sized vector which can then be aggregated into node and edge
features easily and added to the positional embeddings.  This
should create a learning dynamic in which more complex graphs
essentially offer finer control over the interpretation of the
codebook vector, by placing specific dimensions of this codebook
vector in closer/farther proximity to each other.  Simpler graphs,
on the other hand, would offer much more straightforward and
uncomplicated access to the codebook vector itself.