fix bug in duplication of edge features
The edge features were intended to be tiled to create identical features
for bidirectional edges on a per example basis, but were instead being
tiled after the graphs had been batched.  This would have led to a completely
incoherent distribution of the edge features, embedding codebook vectors in
the wrong graphs entirely.