restructure to create symbolically-synthesized neural networks
The previous results were ephemeral, but suggested to me that there was
something to be gained by examining whether a single graph could create
a bias in a neural network that was useful/observable.  I believed it was
possible, after all, that what seemed to be the altered performance of the
experimental models used after all was simply hard to nail down because the
models were required to use so many different graphs.  If instead, they were
required to use only one, it would finally be possible to say, with certainty,
what the effect of this contribution was.  Moreover, the same techniques
applied for maintaining constant population sizes could still be applied.
However they would now be applied based on the training performance of models
structured after individual graphs, rather than based on the usage of a graph
among those selected from by a single model.  This required a thorough
reorganization of the code.